refactor: icons 폴더 위치 변경에 따른 import 경로 수정
